Reviewed all folder structure. fixed all tests names. #142
